==> notebooks/learning_journeys/01_get_gpt_answers.ipynb <==
#Note: The openai-python library support for Azure OpenAI is in preview.
import os

import openai
openai.api_type = "azure"
openai.api_base = "https://andreastestressource.openai.azure.com/"
openai.api_version = "2023-05-15"
openai.api_key =  "1d153d0c7f8b470d856c80e928f94cd4"
deployment_id="AndreasTestBereitstellung"




study_subjects = [
    "Betriebswirtschaftslehre",
    "Informatik",
    "Medizin",
    "Maschinenbau",
    "Elektrotechnik",
    "Psychologie",
    "Jura/Rechtswissenschaften",
    "Architektur",
    "Chemie",
    "Biologie",
    "Geschichte",
    "Soziologie",
    "Volkswirtschaftslehre",
    "Mathematik",
    "Physik",
    "Politikwissenschaft",
    "Medienwissenschaft",
    "Sprachwissenschaft/Linguistik",
    "Pädagogik",
    "Philosophie",
    "Kunstgeschichte",
    "Sportwissenschaft",
    "Geografie",
    "Musikwissenschaft",
    "Ethnologie"
]


lifelong_learning = [
    "Barista",
    "Gärtner",
    "Stricken",
    "Basteln",
    "Malen",
    "Kommunizieren",
    "Meditation",
    "Campingmanagement",
    "Scrum",
    "Azure DevOps",
    "Projektmanagement",
    "Immobilien",
    "Teamleitung"
]


system_role = """You are a helpful assistant who communicates in German and helps to create curriculum in a json format. 

Your answer should always look like this. 

{"title":"Title of the learning goal",
"Skills": [{"name": "skill 1", "description":"Summary of skill 1","why":"why is skill 1 important",  "duration": "X Stunden"}, 
            {"name": "skill 2", "description":"Summary of skill 2","why":"why is skill 2 important","duration": "Y Stunden"}],
"additional info": "everything else you want to say like a description of the learning goal"}

Do not answer anything that is not in that format. """
first_user_input = "Ich möchte ein Experte in %s werden. Sag mir, was ich dazu alles lernen sollte. Liste alle Fähigkeiten in einer sinnvollen Reihenfolge auf. Schreib bei jeder Fähigkeit hinzu, wie viele Stunden ich schätzungsweise brauche, um diese Fähigkeit zu lernen. Ich habe einen Highschool-Abschluss und keine andere Vorerfahrung."
def get_gpt_answer(learning_goal, system_role =system_role, first_user_input=first_user_input):
    try:
        response = openai.ChatCompletion.create(
            deployment_id=deployment_id, # engine = "deployment_name".
            messages=[
                {"role": "system", "content": system_role},
                {"role": "user", "content": first_user_input % learning_goal}
            ]
        )
    except Exception as e:
        print(f"Error occurred: {e}")
        response = f"Error occurred: {e}"

    return response

lifelong_learning_answers = []
for l in lifelong_learning:
    print(l)
    lifelong_learning_answers.append(get_gpt_answer(l))

import json

#save to json
# Combine keys and values into a dictionary
data = dict(zip(lifelong_learning, lifelong_learning_answers))

# Write the data dictionary to a JSON file
with open("data/lifelong_learning_raw.json", "w") as json_file:
    json.dump(data, json_file)

study_subjects_answers = []
for l in study_subjects:
    print(l)
    study_subjects_answers.append(get_gpt_answer(l))

#save to json
# Combine keys and values into a dictionary
data = dict(zip(study_subjects, study_subjects_answers))

# Write the data dictionary to a JSON file
with open("data/study_subjects_raw.json", "w") as json_file:
    json.dump(data, json_file)
