==> notebooks/learning_journeys/02_parse_answers.ipynb <==
import json


with open("data/lifelong_learning_raw.json", "r") as json_file:
    ll_data = json.load(json_file)

with open("data/study_subjects_raw.json", "r") as json_file:
    study_data = json.load(json_file)

ll_data["Barista"]["choices"][0]["message"]["content"]

study_data_parsed = {}
for k in study_data.keys():
    content = json.loads(study_data[k]["choices"][0]["message"]["content"])
    study_data_parsed[k] = content

ll_data_parsed = {}
for k in ll_data.keys():
    content = json.loads(ll_data[k]["choices"][0]["message"]["content"])
    ll_data_parsed[k] = content


study_data_parsed["Betriebswirtschaftslehre"]["Skills"]

with open("data/study_subjects_parsed.json", "w") as json_file:
    json.dump(study_data_parsed, json_file)


with open("data/lifelong_learning_parsed.json", "w") as json_file:
    json.dump(ll_data_parsed, json_file)
