==> notebooks/learning_journeys/03_encode_paths.ipynb <==
import json
from sentence_transformers import SentenceTransformer


with open("data/lifelong_learning_parsed.json", "r") as json_file:
    ll_data = json.load(json_file)

with open("data/study_subjects_parsed.json", "r") as json_file:
    study_data = json.load(json_file)

model = SentenceTransformer('sentence-transformers/paraphrase-multilingual-mpnet-base-v2')


def add_embeddings(data):
    return_dict = {}
    for k in data.keys():
        print(k)
        return_dict[k] = {}
        for skill in data[k]["Skills"]:
            return_dict[k][skill["name"]] = {}
            for skill_k in skill.keys():
                raw_content = skill[skill_k]
                embedding = model.encode(raw_content)
                return_dict[k][skill["name"]][skill_k] = { 
                    "raw_content":raw_content,
                    "embedding":embedding
                }
                    
                

    return return_dict
                
        

ll_data_we = add_embeddings(ll_data)

study_data_we = add_embeddings(study_data)

study_data_we.keys()

ll_data_we["Barista"]["Kaffeebohnen auswählen und mahlen"]

def print_dict_keys(dictionary, indent=''):
    for key, value in dictionary.items():
        if isinstance(value, dict):
            print(f"{indent}{key}:")
            print_dict_keys(value, indent + '  ')
        else:
            print(f"{indent}{key}")

print_dict_keys(ll_data_we)

#now we need pickle because we have numpy data, which is not serializable
import pickle



with open("data/study_subjects_with_embeddings.pkl", "wb") as pkl_file:
    pickle.dump(study_data_we, pkl_file)


with open("data/lifelong_learning_with_embeddings.pkl", "wb") as pkl_file:
    pickle.dump(ll_data_we, pkl_file)
