==> notebooks/learning_journeys/04_cos_similarity.ipynb <==
### Notebook to create learning journeys (recommendations of videos to watch to learn a goal) by using encodings and cos similarity

import json
from sentence_transformers import SentenceTransformer
import pandas as pd
import numpy as np

model = SentenceTransformer('sentence-transformers/paraphrase-multilingual-mpnet-base-v2')

def get_data(path): 
        merged_df = pd.read_csv(path)
        strings = merged_df['String']
        str_lst = strings.values

        vocab = merged_df['Title'].values
        identifier = merged_df['identifier']
        identifier_vocab = pd.DataFrame({'ID': identifier, 'Vocab': vocab})
        identifier_vocab = identifier_vocab.set_index('Vocab')['ID'].to_dict()
        return merged_df, str_lst, vocab, identifier_vocab, identifier


merged_df, str_lst, vocab, identifier_vocab, identifier = get_data('data/merged_data_for_AI.csv')


import pickle


with open("data/study_subjects_with_embeddings.pkl", "rb") as pkl_file:
    study_data_we = pickle.load(pkl_file)


with open("data/lifelong_learning_with_embeddings.pkl", "rb") as pkl_file:
    ll_data_we = pickle.load(pkl_file)


with open("data/title_encoded.pkl", "rb") as pkl_file:
    title_encoded = pickle.load(pkl_file)

video_embeddings = pd.read_csv('data/sentence_transformer_embedding.csv')
del video_embeddings['Unnamed: 0']
video_embeddings_values = video_embeddings.values
print(video_embeddings_values.shape)

from sklearn.metrics.pairwise import cosine_similarity

#function to calculate the cosine similarity between an matrix A and a vector b. Matrix A is the document database and vector b the search query. 
def calculate_cosine_similarity(A, b):
    # Convert A and b to numpy arrays
    A = np.array(A)
    b = np.array(b)

    # Calculate cosine similarity between A and b
    similarity_scores = cosine_similarity(A, b.reshape(1, -1)).flatten()

    # Sort the indices based on similarity scores
    sorted_indices = np.argsort(similarity_scores)[::-1]

    return sorted_indices


def get_top_n_videos(subject,lj_data, database_embeddings, video_titles, video_identifier, n = 3):
    '''
    function to get the top n videos with the highest cos similarity

            Parameters:
                    subject (str): The name of the subject someone wants to study. e.g. "Informatik"
                    lj_data (dict): A dict with different learning journeys and their specific steps. Al steps are NLU encoded. 
                    database_embeddings (dict): A numpy array with the embeddings of the documents. 
                    video_titles (list of str): The titles of the videos, to print them for the relevant videos. 
                    video_identifier (list of str): The ids of the youtube videos, to print the link to the video. 
                    n (int): number of videos that should be recommended per learning step

            
    '''
    subject_dict = lj_data[subject]
    print("Learning journey for ", subject, " is:")
    for i in subject_dict.keys():
        
        print(i)
        tmp_string = subject_dict[i]["name"]["raw_content"]
        t_emb = subject_dict[i]["name"]["embedding"]
        print("to learn ", tmp_string, " watch the following videos:")
        cosine_similarity_sorted = calculate_cosine_similarity(database_embeddings,t_emb)
        for index, v in enumerate(cosine_similarity_sorted[:3]):
            title = video_titles[v]
            #print(index)
            #print(title)
            url = "https://www.youtube.com/watch?v=" + video_identifier[v]
            print("Video %i: %s" % (index + 1, title))
            print(url)
        print("____________________________" + "\n")


### The learning journeys are stored in the dictonaries study_data_we and ll_data_we. The keys are the subjects they contain journeys for

study_data_we.keys()

ll_data_we.keys()

### title_encoded and video_embeddings_values are both valid encoding databases. 
- title_encoded: only the title of the videos are encoded. 
- video_embeddings_values: title, uploader and the transcript are concatenated. Of this concatenation the first 128 tokens are encoded. 

Change the database to the variable you want to use for the encoding database

database = video_embeddings_values
#database = title_encoded

get_top_n_videos("Informatik", study_data_we, database,merged_df["Title"],merged_df["identifier"])
