==> notebooks/learning_journeys/05a_inspect_corrected_transcripts.ipynb <==
### A short notebook to inspect the results of the punctuation correction

import numpy as np
import pickle
import pandas as pd

import textwrap

def pprint(long_string, width = 1150):
    # Use the textwrap module to pretty print the string
    pretty_string = textwrap.fill(long_string, width=width)
    print(pretty_string)


with open("data/punctuation_correction/corrected_transcripts_all.pkl", "rb") as file:
    co_transcripts = pickle.load(file)


def get_data(path): 
        merged_df = pd.read_csv(path)
        strings = merged_df['String']
        str_lst = strings.values

        vocab = merged_df['Title'].values
        identifier = merged_df['identifier']
        identifier_vocab = pd.DataFrame({'ID': identifier, 'Vocab': vocab})
        identifier_vocab = identifier_vocab.set_index('Vocab')['ID'].to_dict()
        return merged_df, str_lst, vocab, identifier_vocab, identifier


merged_df, str_lst, vocab, identifier_vocab, identifier = get_data('data/merged_data_for_AI.csv')

merged_df["co_transcripts"] = co_transcripts

### Lets compare the number of sentences in the original transcripts and in the corrected versions of them. We use the sumy tokenizer to seperate the strings into sentences. 

from sumy.nlp.tokenizers import Tokenizer as SumyTokenizer
sumy_tokenizer = SumyTokenizer("de")


def count_sentences(text):
    global sumy_tokenizer
    try:  
        sentences =  sumy_tokenizer.to_sentences(text)
        sentence_count = len(sentences)
    except Exception:
        sentence_count = np.nan
    
    return sentence_count


merged_df["#sentences_original"] = merged_df["transcript"].apply(count_sentences)
merged_df["#senteces_corrected"] = merged_df["co_transcripts"].apply(count_sentences)

diff = merged_df["#senteces_corrected"]/ merged_df["#sentences_original"]
diff = [x for x in diff if not (np.isnan(x) or np.isinf(x))]
print(np.mean(diff))

### the corrected texts have nealry 15 times the amount of sentences! Lets check in an example, how the original version and the corrected one differ. 

#index that is used. Change, to see another example
example_index = 0

#helper function that prints the text in a readable output and also the amount of sentences. 
def analyze_text(text):
    print("# Sentences: ", count_sentences(text))
    pprint(text)

analyze_text(co_transcripts[example_index])

analyze_text(merged_df["transcript"][example_index])
