==> notebooks/learning_journeys/99_count_tokens.ipynb <==
import json
from sentence_transformers import SentenceTransformer
import pandas as pd

model = SentenceTransformer('sentence-transformers/paraphrase-multilingual-mpnet-base-v2')

def get_data(path): 
        merged_df = pd.read_csv(path)
        strings = merged_df['String']
        str_lst = strings.values

        vocab = merged_df['Title'].values
        identifier = merged_df['identifier']
        identifier_vocab = pd.DataFrame({'ID': identifier, 'Vocab': vocab})
        identifier_vocab = identifier_vocab.set_index('Vocab')['ID'].to_dict()
        return merged_df, str_lst, vocab, identifier_vocab, identifier


merged_df, str_lst, vocab, identifier_vocab, identifier = get_data('data/merged_data_for_AI.csv')


from transformers import BertTokenizer, TFBertModel
#we assume the tokenizer of different Bert models is similar or the same
tokenizer = BertTokenizer.from_pretrained('bert-base-cased')

def print_first_25_words(strings):
    for string in strings:
        
        # Split the string into words
        words = string.split()

        # Print the first 50 words
        first_25_words = " ".join(words[:50])
        print(first_25_words)
        tokenized_input = tokenizer.tokenize(first_25_words)
        print("# tokens: ", len(tokenized_input))
        print("_________________________________")

print_first_25_words(str_lst[:100])

full_string = str_lst[0]
print(full_string)
half_string = full_string[:int(len(full_string)/2)]
quarter_string = full_string[:int(len(full_string)/4)]


enc_full_string = model.encode(full_string)
enc_half_string = model.encode(half_string)
enc_quarter_string = model.encode(quarter_string)

import numpy as np

print(np.array_equal(enc_full_string, enc_half_string))
print(np.array_equal(enc_full_string, enc_quarter_string))

index = 0
for index in range(10):
    print("characters: ", len(str_lst[index]))
    tokenized_input = tokenizer.tokenize(str_lst[index])
    print("# tokens: " , len(tokenized_input))

a = model.encode(str_lst[:200])
a.shape

import textwrap

def pprint(long_string, width = 40):
    # Use the textwrap module to pretty print the string
    pretty_string = textwrap.fill(long_string, width=width)
    print(pretty_string)


pprint(str_lst[0])

video_embeddings = pd.read_csv('data/sentence_transformer_embedding.csv')
del video_embeddings['Unnamed: 0']
video_embeddings_values = video_embeddings.values
print(video_embeddings_values.shape)

video_embeddings.columns

#now we need pickle because we have numpy data, which is not serializable
import pickle



with open("data/study_subjects_with_embeddings.pkl", "wb") as pkl_file:
    pickle.dump(study_data_we, pkl_file)


with open("data/lifelong_learning_with_embeddings.pkl", "wb") as pkl_file:
    pickle.dump(ll_data_we, pkl_file)
